requirement.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.0.0
Django==1.11
django-braces==1.8.1
django-environ==0.4.0
djangorestframework==3.7.0
kombu==4.1.0
psycopg2==2.6.2
py==1.5.2
pytest==3.0.6
pytest-django==3.1.2
pytz==2017.3
six==1.11.0
vine==1.1.4
wheel==0.24.0
